requirements.txt
praw==7.7.1
python-dotenv==0.19.2
rich==13.7.1
requests==2.27.1
pyahocorasick==2.1.0